Fix some type errors when running TS type check for win32 platform
