+++ apps/fluent-tester/src/TestComponents/Common/MenuPicker.tsx
@@ -1,26 +1,13 @@
 import * as React from 'react';
 import { Picker } from '@react-native-picker/picker';
 import { testProps } from './TestProps';
+import type { CollectionItem, MenuPickerProps } from './MenuPicker.types';
+export { CollectionItem, MenuPickerProps };
 
 /*
  * The MenuPicker was created because the RN Core Picker was deprecated (preventing us from updating to RN 0.66).
  * MenuPicker uses the community Picker package for iOS/Win32 and our own Picker for Win32/MacOS.
  */
-
-export interface CollectionItem<T = string> {
-  label: string;
-  value?: T;
-  testID?: string;
-}
-
-export interface MenuPickerProps {
-  prompt?: string;
-  selected?: string;
-  onChange?: (value: any, index?: number) => void;
-  collection?: CollectionItem[];
-  style?: any;
-  testID?: string;
-}
 
 export const MenuPicker: React.FunctionComponent<MenuPickerProps> = (props: MenuPickerProps) => {
   const { prompt, selected, onChange, collection, style, testID } = props;
