+++ packages/experimental/Shimmer/.eslintrc.js
@@ -1,0 +1,3 @@
+module.exports = {
+  extends: ['@fluentui-react-native/eslint-config-rules'],
+};
