+++ packages/experimental/Shimmer/src/Shimmer.tsx
@@ -1,8 +1,11 @@
 /** @jsx withSlots */
 import { useRef, useEffect, useMemo, useCallback } from 'react';
-import { Circle, ClipPath, Defs, LinearGradient, Rect, Stop, Svg, G, TransformObject } from 'react-native-svg';
-import { shimmerName, ShimmerProps, ShimmerType } from './Shimmer.types';
-import { compose, mergeProps, withSlots, UseSlots, buildUseStyling } from '@fluentui-react-native/framework';
+import type { TransformObject } from 'react-native-svg';
+import { Circle, ClipPath, Defs, LinearGradient, Rect, Stop, Svg, G } from 'react-native-svg';
+import type { ShimmerProps, ShimmerType } from './Shimmer.types';
+import { shimmerName } from './Shimmer.types';
+import type { UseSlots } from '@fluentui-react-native/framework';
+import { compose, mergeProps, withSlots, buildUseStyling } from '@fluentui-react-native/framework';
 import { Animated, I18nManager } from 'react-native';
 import { stylingSettings } from './Shimmer.styling';
 import assertNever from 'assert-never';
