+++ packages/experimental/Shimmer/src/Shimmer.types.win32.ts
@@ -1,5 +1,7 @@
-import { ViewProps } from 'react-native';
-import { ShimmerProps, ShimmerTokens } from './Shimmer.types';
+import type { ViewProps } from 'react-native';
+import type { ShimmerProps, ShimmerTokens } from './Shimmer.types.shared';
+export { ShimmerProps, ShimmerTokens };
+export { shimmerName, ShimmerRectElement, ShimmerElementTypes, ShimmerCircleElement } from './Shimmer.types.shared';
 
 export interface ClippingMaskProps extends ShimmerTokens {
   /**
