+++ packages/experimental/Shimmer/src/Shimmer.win32.tsx
@@ -1,10 +1,14 @@
 /** @jsx withSlots */
-import { compose, mergeProps, UseSlots } from '@fluentui-react-native/framework';
+import type { UseSlots } from '@fluentui-react-native/framework';
+import { compose, mergeProps } from '@fluentui-react-native/framework';
 import { View } from 'react-native';
-import { ClipPath, Defs, LinearGradient, Path, Rect, Stop, Svg, SvgProps } from 'react-native-svg';
+import type { SvgProps } from 'react-native-svg';
+import { ClipPath, Defs, LinearGradient, Path, Rect, Stop, Svg } from 'react-native-svg';
 import { stylingSettings } from './Shimmer.styling.win32';
-import { ShimmerElementTypes, shimmerName, ShimmerProps } from './Shimmer.types';
-import { ClippingMaskProps, ShimmerType, ShimmerWaveProps } from './Shimmer.types.win32';
+import type { ShimmerElementTypes, ShimmerProps, ShimmerCircleElement, ShimmerRectElement } from './Shimmer.types.shared';
+export { ShimmerCircleElement, ShimmerRectElement };
+import { shimmerName } from './Shimmer.types.shared';
+import type { ClippingMaskProps, ShimmerType, ShimmerWaveProps } from './Shimmer.types.win32';
 import { RCTNativeAnimatedShimmer } from './consts.win32';
 import { convertRectToSvgPath, convertCircleToSvgPath } from './SvgShapeToPath';
 import { withSlots } from '@fluentui-react-native/framework';
