+++ packages/experimental/Shimmer/src/index.ts
@@ -1,2 +1,11 @@
-export * from './Shimmer';
-export * from './Shimmer.types';
+export { Shimmer } from './Shimmer';
+export { shimmerName } from './Shimmer.types';
+export type {
+  ShimmerCircleElement,
+  ShimmerElementTypes,
+  ShimmerProps,
+  ShimmerRectElement,
+  ShimmerSlotProps,
+  ShimmerTokens,
+  ShimmerType,
+} from './Shimmer.types';
